requirements.txt
# File: requirements.txt

requests
Pillow